+++ src/Dax.Template/Constants/Attributes.cs
@@ -10,4 +10,5 @@ public static class Attributes
     public const string SqlbiTemplateTableDateAutoTemplate = "DateAutoTemplate";
     public const string SqlbiTemplateTableHolidays = "Holidays";
     public const string SqlbiTemplateTableHolidaysDefinition = "HolidaysDefinition";
+    public const string SqlbiTemplateTableCalculationGroup = "CalculationGroup";
 }